Fashion Project Flask/requirements.txt
Flask
numpy
tensorflow
pandas
torch
clip-anytorch
Pillow
pickle-mixin
scipy
scikit-learn
requests
Werkzeug
